refactor: keep ActivityIndicator in components folder
